# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
    personId:
      name: personId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Person identifier
...
  schemas:
...
    PersonSex:
      type: string
      enum: [MALE, FEMALE, OTHER, UNKNOWN]
    PrivacyLevel:
      type: string
      enum: [PUBLIC, FAMILY, PRIVATE]
...
    ParentageType:
      type: string
      enum: [BIOLOGICAL, ADOPTIVE, STEP, FOSTER, UNKNOWN]
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    TokenPair:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          example: 900
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── User ────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isVerified, plan, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          example: "en"
        timezone:
          type: string
          example: "America/New_York"
        isVerified:
          type: boolean
        plan:
          type: string
          enum: [FREE, BASIC, PREMIUM, FAMILY, PROFESSIONAL]
        createdAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          pattern: "^[a-z]{2}(-[A-Z]{2})?$"
        timezone:
          type: string

    # ── Tree ────────────────────────────────────────────────
...
    PersonSummary:
      type: object
      description: Lightweight representation used in list responses and tree views
      required: [id, sex, isLiving, isDeceased]
      properties:
        id:
          type: string
          format: uuid
        sex:
          $ref: "#/components/schemas/PersonSex"
        isLiving:
          type: boolean
        isDeceased:
          type: boolean
        displayGivenName:
          type: string
          nullable: true
        displaySurname:
          type: string
          nullable: true
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
        birthYear:
          type: integer
          nullable: true
          description: Derived from BIRTH event; null if unknown
        deathYear:
          type: integer
          nullable: true
...
    AncestorNode:
      type: object
      required: [person, depth, familyGroupId]
      properties:
        person:
          $ref: "#/components/schemas/PersonSummary"
        depth:
          type: integer
          description: "1 = parent, 2 = grandparent, etc."
        familyGroupId:
          type: string
          format: uuid
        parentageType:
          $ref: "#/components/schemas/ParentageType"

    # ── Family Group ─────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getMe
      summary: Get the current user's profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/401Unauthorized"

    patch:
      tags: [Users]
      operationId: updateMe
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"

    delete:
      tags: [Users]
      operationId: deleteMe
      summary: Delete the current user's account (scheduled deletion — 30-day grace period)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmPassword]
              properties:
                confirmPassword:
                  type: string
      responses:
        "202":
          description: Deletion scheduled — 30-day grace period before permanent removal
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
  /users/me/change-password:
    post:
      tags: [Users]
      operationId: changePassword
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 10
                  maxLength: 128
      responses:
        "200":
          description: Password changed — all existing sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
...
  /trees/{treeId}/persons/{personId}/ancestors:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: getAncestors
      summary: Get all ancestors of a person (unlimited generations)
      parameters:
        - name: maxDepth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: parentageTypes
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ParentageType"
          style: form
          explode: false
          description: "Filter by parentage type (default: all types)"
      responses:
        "200":
          description: Ancestor list ordered by depth
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AncestorNode"
                  subject:
                    $ref: "#/components/schemas/PersonSummary"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /trees/{treeId}/persons/{personId}/descendants:
    parameters:
      - $ref: "#/components/parameters/treeId"
      - $ref: "#/components/parameters/personId"
    get:
      tags: [Persons]
      operationId: getDescendants
      summary: Get all descendants of a person (unlimited generations)
      parameters:
        - name: maxDepth
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: parentageTypes
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ParentageType"
          style: form
          explode: false
      responses:
        "200":
          description: Descendant list ordered by depth
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AncestorNode"
                  subject:
                    $ref: "#/components/schemas/PersonSummary"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...